Add comment string and formatoptions to Matlab filetype
These are not set in the ftplugin file, hence when changing the filetype
from a different one might be missing or wrong. This happens for example
in Firenvim where the filetype is set to Vimwiki per default but might
be changed to Matlab when editing comments in Gitlab containing Matlab
code.